Pipeline and Splitjoin now nonfinal
Programmers will want to subclass Pipeline to create named
Pipelines for reuse at multiple places in the stream graph, and
similarly with Splitjoins.  Note that all the subclass can do is
pass arguments to the superclass constructor; no methods can
be overridden.

This isn't the best way to do this.  Because super() must come
as the first statement in the subclass constructor, many subclasses
will use super(someStaticMethodToCreateInstances(args)), which is
unnecessary boilerplate.  Perhaps switching to an add() method
would be better than using constructor parameters.