Introduce PrimitiveWorker to support interpreter
PrimitiveWorker is a package-private class that provides plumbing
for the interpreter.  Using a class allows this plumbing to be
reused in multiple places (Filter, Splitter and Joiner) rather than
repeated in each.  Making it package-private ensures user code
can't interfere with it (and users can't get confused by it), but
has the downside of requiring the interpreter to be in the same
package even though it's logically seperate.

Another option would be to have PrimitiveWorker encapsulate a
worker class instance, but supporting outputs() and inputs()
requires the worker class to keep a reference to the PrimitiveWorker
encapsulating it, largely destroying any decoupling benefits such
a scheme would provide.

I've probably made some typos in a commit this large; sorry.