Introduce OneToOneElement interface
This makes Splitters and Joiners StreamElements without allowing
them to be put in arbitrary places in the graph.

This may be a mistake, though I would point out that StreamIt (the
language) has the concept of a one-to-one element, it's just left
implicit in the definition of the "add" statement.  We need to be
explicit here.

Perhaps a more concise name than OneToOneElement would make this
feel more natural?