Remove copy() from StreamElement interface
The intent behind the copy() method was to allow StreamCompilers to
make copies of the user-provided stream graph passed to compile(),
so that the user could not accidentally (or intentionally) interfere
with the compiler by mutating the stream graph.  Additionally, users
may have found it useful for creating "prototype" stream graphs that
could be further modified.

Unfortunately, Portal depends on reference equality of workers that
subscribe to messages, and the error resilience provided by copy()
is not worth the effort of fixing up the graph to point to new
objects, especially when final fields are involved.  Thus copy() has
been removed from StreamElement.