Portal.Handle: don't remember latency
We assume in MessageConstraint that the latency is a constant, so
we don't need to remember it in the Handle (indeed, it's a dead
field, which is how I realized this).  We'd need to reintroduce it
if we ever support variable latency, e.g. "latency i where i will
always be in [2, 5]" or whatever.  Unrestricted latency is also
technically possible for downstream messages but let's not go there.