Klass: initialize other members earlier
See the previous comment for why; we might as well do the name and
modifiers as well as they'll never cause mutual recursion problems.

It'd be really nice for the superclass to be initialized, but that
might be mutually recursive (through the superclass fields or
methods), so we need to be in the Module before then.