MethodResolver: void returns return void
Even if we're a constructor, we don't really have a value to return.
We could fake returning uninitializedThis, but there's not much point
because we'd just undo that transformation when emitting bytecode.
It would only matter if we decide to inline constructors, and they're
special enough that either way will require special handling.