MethodResolver: ensure block only visited once
In the old version, a block was only added to the visited set after
being removed from the worklist, so if it was added to the worklist
twice, it would be visited twice.  We now add to the visited set as
soon as the block is put on the worklist.