User: don't call subclass checkOperand for nulls
This forces all subclasses to accept null operands, required by
dropAllOperands().  We could also just write the subclasses to be
null-tolerant, but that's more work.

Another option would be to just have dropAllOperands() remove all
our Use objects to permit GC.  This doesn't require messing with
checkOperand at all, but would have the side effect of changing the
number of operands of a dropped-operand user.  Perhaps this is even
a feature, if we consider dropAllOperands() to be a destructor, in
that we'll fail more quickly if we try to use such a destroyed
object.  Explicit destructor functions (probably named close()) would
likely be better, however.