Buffer cannot be generic
In the runtime, we'll never know the type, so we'll always have
Buffer<?>.  Java's loose array type checking allows
write(Object[],int,int) to work, but write(T) never works for a
Buffer<?> because we don't know the type.

As there's no benefit because we'll always use Buffer<?> anyway,
so we'll not be generic.

This also avoids a situation similar to Worker<?, ?> where we have
lots of unchecked casts to please the compiler.