BlobHostStreamCompiler: a StreamCompiler delegating to a Blob
DebugStreamCompiler could be refactored to use this, maybe.

BlobHostStreamCompiler uses ConcurrentArrayBuffer, which uses
ConcurrentPeekableQueue, which seems to have some (more) bugs.
In this case they went away when I changed the assertion in
CAB.readAll() to provide the offset in the exception message, and
when I removed that change the bug didn't come back.  Just something
to check if you see more bugs here.