MethodResolver: check for blocks without terminators; Compiler fixes
Without a terminator, phi moves don't get inserted.  Compiler only
left the terminators out in straight-line code so it didn't matter,
but we should check to be sure.  (Note that fallthroughs are not
necessarily safe if they're one edge of a control-flow merge; phi
moves wouldn't get generated for any definitions in the block.)