Schedule.getSchedule() for backcompat
In the future I'd like Schedule to absorb other functions, such as
computing the required buffer size or listing the possible
topological sorts of filter ordering dependencies.  But for now we
just extract the execution counts so we can plug it in where we
would have used Scheduler.schedule()'s return value.