Compiler: size buffers considering init schedule
This allows us to handle a round-robin splitjoin with Identity on
one branch and Identity-Identity-Identity on the other: we buffer
more than the steady-state requirement so we can find a valid init
schedule.

This passes some StreamFuzzer fuzz instances with splitjoins, but
others fail with infeasible external schedules.  I'm not sure if
this is a compiler, ILP solver interface, or fuzzer bug.