Turn splitjoins back on in StreamFuzzer
I think the "system is infeasible" exceptions are due to rate
explosion -- usually the debug compiler can't produce any output
with 10000 inputs.  If I add larger rates on the Batchers and use
those exclusively, I can get the error at smaller splitjoin depths.
But it's weird to hit this so soon -- I wonder if we should consider
cyclo-static dataflow for splitters and joiners to keep rates lower?