api.Input/Output and related changes
This change introduces Input and Output classes in the API.
CompiledStream.offer() and poll() have moved to ManualInput and
ManualOutput respectively.  Other Input/Output implementations
read from files, collections, or other structures; thus this finally
implements a FileReader equivalent.  This allows the stream to pull
input, rather than requiring user threads to push input in and take
input out, which should improve efficiency.

The non-manual inputs immediately put the stream in the "draining"
state, so when they run out, the stream will clean itself up.
ManualInput provides a drain() method (which should probably be named
close() to match Java convention) to initiate draining.

CompiledStream once again provides methods to wait for shutdown,
now called awaitDrained() (instead of awaitDraining()).  We removed
these when we introduced Blobs and Buffers.  Unfortunately, I did
not record why.  Now that most uses don't have to poll for output,
we really want a way to find out when the stream is complete other
than polling isDrained(), and awaitDrained() is easily implemented
via Blob.drain() callbacks.  A future extension might allow registering
callbacks or providing Future instances to block on, for better
integration with other concurrent code.

DebugStreamCompiler is now a BlobHostStreamCompiler using a
DebugBlobFactory, which means it runs on its own thread instead of
borrowing the offer()-ing thread.  Obviously, many streams won't
have an offer(), so we have to do something.  We could add a
doSomeWork() method to DebugStreamCompiler that the user could call,
but that would prevent using DebugStreamCompiler as a drop-in
replacement for the other compilers when debugging is required.

Only the BlobHostStreamCompiler is updated in this commit (and the
DebugStreamCompiler is now a BlobHostStreamCompiler).

Many of the main() methods in the app classes have been commented out
rather than updated.  The preferred way to run individual benchmarks
is through the Benchmarker, which will soon become a non-static class.

Many of the apps will now want to import real test data from the
classic StreamIt repository.