StreamVisitor: maintain depth after skipped traversals
Before this change, if we skipped any sections of the graph by
returning false from enterPipeline or enterSplitjoin, depth was not
decremented, so endVisit would never be called.  This caused
SerialCompositeStreamVisitor to never call its component visitors,
because it does so in endVisit, which was not called.  This meant
CheckVisitor didn't actually check anything.

Aside: I built CheckVisitor specifically for the splitjoin balance
check, because the StreamFuzzer was generating graphs the compiler
couldn't schedule and I wasn't sure if it was rate blowup or a fuzzer
problem.  Upon seeing CheckVisitor not raise any errors, I assumed
rate blowup.  Well, now that CheckVisitor is checking, the fuzzer is
constantly raising errors.

I even said to Sumanan while pair programming that I was only assuming
CheckVisitor worked because I'd never seen it fail and I didn't have
any tests.