Remove NestedSplitjoinExample
This is a useful thing to have in a test, but I think it's a bit too
complicated to work out for a human to be a sanity test.  I could
enhance the splitjoin simulator from RoundrobinSanity for this, but
at some point you're just reimplementing an interpreter.

I think these cases are best handled with a regression test that
compares against the DebugStreamCompiler output.  This specific case
is well-covered by StreamFuzzer.