Dataset: just use Input<?>
We're already erasing all the types to Object, so we might as well
do it in one place.  There is a "proper" way to do this: Benchmark<I,O>
and Dataset<I,O>.  The types get erased, but we can capture the wildcards
in the Benchmarker so that the compiler knows the ? stand for the same
types.  That just seems like a lot of work for not much gain, though.