Compiler: fusion, sort of
This implements fusion in the sense of fusing StreamNodes.  Actual
worker work methods are not fused.  Mostly this just removes buffering,
which speeds us up because we fit in cache better.  Theoretically, a JVM
could use inlining to do actual fusion at this point, but our index
computations make such heavy use of arrays that I wouldn't expect it to.

Currently the default is to fuse whenever possible.  Obviously the
autotuner can make other choices but it might be nice to try out some
heuristics, at least for testing purposes. (use the fuzzer for random
fusion?)

By the way, we now get speedups from multiple cores on Windows.  Maybe it
was just a cache issue, or something.