Disable DeadCodeElimination.eliminateTriviallyDeadInsts()
I've had problems with this one before, but the previous fix was woefully
inadequate.  Just because the result of a pop() call isn't used doesn't
mean pop() is dead, because it has side effects!

At some point there will be a side-effects analysis; for now we'll just
turn this off.