Create global storage
I'm still confused about the capacity of ConcreteStorage objects and how
they will be initialized (in particular, how to deal with buffered data,
either generated for peeking, carried from DrainData initial state or as a
result of rate-lag between branches of a splitjoin).  Still, I can get
some of the other scaffolding working in the meantime.