Memory leak fix
The ObjectOutputStream maintains a table mapping objects written into the
stream to a table.This table maintains references to objects that might
otherwise be unreachable by an application, thus, resulting in an unexpected
situation of running out of memory. A call to the ObjectOutputStream.reset()
method resets the table to its initial state, allowing all previously written
objects to be elgible for garbage collection.