Remove println in Compiler2BlobHost
There's a GC-dependent crash in JDK 7u45, possibly with MethodHandle
involvement.  If the GC is triggered during the phaser's onAdvance
method, the JDK segfaults.  This is a timing-dependent bug but
reproduces fairly reliably on FMRadio with enough input (>4000ish
items).  Usually the process just hangs but the G1 collector
(-XX:+UseG1GC) crashes cleanly with a dump.  For now I've reverted to
7u25.