ActorGroup.specialize: don't assume 1-input/output actors are filters
A RoundrobinSplitter with one output is still a splitter and uses the
splitter archetype signature.  For the indices we inspect the
MethodHandle we're binding to in case we ever spin specialized code
sequences for this case.  For the read/write handles we have to check
the worker because the MethodHandle argument doesn't encode its argument
types, just that it's a MethodHandle of some kind.