Assume less when computing actualInit
This bug was exposed by splitter removal, which creates index functions
with gaps.

I'm not sure about this solution either, particularly as to why we have to
decide on stopping before removing the writes from required reads (swap
those two lines and we're off by one on most token edges).