ActorGroup.inputs()/outputs(): now based only on internal/external
If we read from an external storage, it's an input; if we write into an
external storage, it's an output.  This might mean a storage is
simultaneously an input and output of the same ActorGroup, but that's just
how it is.  Splitter removal can create storages where e.g. the only
input and outputs 0 and 2 are in group A and output 1 is in group B, which
previously caused us to not consider the storage an input to group A.