ActorGroup.specialize(): use loops, not lots of handles
Previously we generated one handle per group iteration, and for splitters
and joiners, one handle per subiteration.  This consumed a lot of memory.
We now use loops keying on the iteration number.

We now depend on modifications to the index arrays persisting through the
steady-state iteration, so we no longer copy them in work methods.
Instead we reinitialize them at the beginning of the ActorGroup outer
loop, once per allocation invokation.