Compute init schedule using the ILP solver
I convinced myself I couldn't use the ILP solver after removals, because
storage was no longer one-to-one.  Consider a graph with two Identity
filters writing to a storage where we need four elements (as the result of
a roundrobin splitter elimination): there are five ways to satisfy the
requirement, but only the (2,2) combination is correct (the rest leave
holes in indices we read from).

I wrote some rather convoluted code to find the init schedule after
removals, but because it was after removals, it could schedule
"half-iterations" of the removed actors.  As we rely on the interpreter to
drain, this resulted in graphs that would run fine in the steady state but
produce incorrect results during draining because a roundrobin splitter's
parity was incorrect.  (See the regression test I recently committed.)

However, Sumanan and I convinced ourselves that we can safely schedule the
initial graph for init, then perform removals.  We might buffer more
elements than strictly necessary, but doing so is safe.  Importantly, the
constraints include splitters and joiners, so we'll only schedule atomic
iterations of them, avoiding our previous problem.

This doesn't pass all the regression tests and we still fail when removing
splitters after fusion, but it's at least more correct and less
complicated.