Compiler2: remove the offset, correctly compute SS capacity
In the most recent regression test, when both fusion and splitter removal
are enabled, we previously miscomputed the steady-state capacity as 5 when
it's actually 6 because we didn't consider the indices being written that
iteration.  This change fixes that.

This change also removes the offset; we no longer try to align the indices
live after initialization to the expected steady-state reads.  This makes
the compiler a bit simpler, but does mean future double-buffered
ConcreteStorage implementations will have to account for the offset
themselves.  Also, whereas with the offset some of the steady-state index
function adjustments were no-ops and could be omitted (though we didn't
check for that), without the offset all adjustments matter and must be
performed, costing an extra addition for the itemsWritten or itemsRead.  I
might come back to this in the future, but for now I'll assume an extra
addition isn't a big deal.

Some of the other regression tests are still failing.