Attempt at drain instruction changes (wrong)
This code attempts to create drain instructions that put some of the live
items as inputs to readers and others as outputs of writers.  This code
does not work; the approach itself is unsound as it won't ever put items
on the middle edge(s) of a nested removal, thus necessarily getting some
items out of order in the nested case.

This commit will be reverted.