Ignore nbproject/configs
These are configurations that can be used to run the project.  (I believe
this was inspired by a similar Eclipse feature.)  I use this for passing
different sets of arguments to Benchmarker to run different sets of tests,
but NetBeans doesn't save the command-line args in the configs (presumably
it assumes they might have system- or user-specific paths, usernames,
passwords etc.), so keeping them version-controlled is useless.