Storage.computeSteadyStateRequirements doesn't need live index set
The first read index is always live or written this iteration, and there
can't be anything live before that, so we can replace the first live index
with the first read index.

The last read index is always either live or less than the last write
index; if it weren't, it would get shifted by throughput at the end of the
iteration and thus not be live in the next iteration.  So we can replace
the last live index with the last read index.