Unbox storage (not work methods)
With this commit, we unbox storage whenever we know it holds only wrapper
instances, using asType to adapt the read and write method handles to
return the archetypal input and output types (always object types).  This
seems to slow us down by introducing extra boxing and unboxing, but may
still be a net win when using large multipliers because we can fit more
primitives than wrappers in the cache.