BulkReadInstruction
This removes a copy: instead of readAll -> array in ReadInstruction ->
ConcreteStorage, we let the ConcreteStorage read into itself.  It won't
usually be a System.arraycopy like you think (but an array-backed Buffer
could), but just saving the copy is valuable enough.

Note there are conditions on this: only if we didn't unbox the input
storage and the indices are contiguously increasing (i.e., we didn't
remove a joiner immediately before the input).  And of course the
ConcreteStorage has to support it.