Arrayify index functions
This uses a generic IndexFunctionTransformer interface, so we can do other
kinds of transformations, such as searching for arithmetic simplifications
(in the event the JVM can't figure them out due to not knowing the
function's exact domain).

This is a compile time regression because we evaulate the domains even if
we're using the identity transformer (which doesn't look at it).  We could
recover this by making that parameter a Callable<NavigableSet<Integer>>
instead.  (The arrayifying transformers need the domain, so we can't get
away with not computing it.)