Bytecodifier.bytecodify
This is part of a plan to add bytecode hooks into the compiler's
MethodHandle chain to help the JVM understand.  In particular, the JVM
doesn't seem to count backedges in our worker loops, so we need to turn
the CompileThreshold parameter way down for it to compile our methods.
Hopefully it'll track bytecode-level invokes.