Compiler2BlobHost: fix the overall loop
For migration, our live items calculation is correct.  For draining, we
drained after executing the core code but before reads and writes, so
we've consumed and produced an execution's worth of items relative to the
live items after-init calculation.  This caused us to lose some outputs
but retain the corresponding inputs, which is correct for stateless graphs
(if wasteful) but incorrect for stateful ones.

To address this, I changed the loop structure.  We now do:
--init and migrate as before
--read, or drain
  LOOP:
--core code
--write outputs
--adjust buffers
--read, or drain
--goto LOOP

This ensures that we're always at our after-init state when we drain
(except for reads that succeeded before reads that failed, but
ReadInstruction.unload() addresses that).  It does mean we'll sometimes do
a bunch of init work but not take a steady-state, but that's a necessary
tradeoff for correctness.