Remove things from Bytecodifier.BOUNCER
I used to have an assert that checked the item was removed, but the
newly-spun class wasn't necessarily initialized yet, so it sometimes
failed.  I removed it with a comment like "I trust CHM.remove() to do what
it says on the tin"... but I wasn't calling it.

This was a major memory leak that eventually ran the fuzzer out of memory.