New core allocation parameterization
Each group and core gets a float in [0,1].  When we allocate a group, we
give each core iterations proportional to its share parameter, in
decreasing order of share, until no iterations remain.  Unallocated
iterations are spread equally.

Share parameters are fractional values, so they don't depend on the
multiplier the way the old iteration-based parameters did.  Sorting by
share avoids the "always core 0" problem as well; now core N has an equal
chance of receiving the largest allocation as core 0.  However, it's still
difficult to get an even data-parallelization across all cores (requires
all cores to have equal share), and it's newly difficult to simply assign
all iterations to a single core (requires that core to have a share of 1).

The default configuration equally data-parallelizes, so we can remove code
to set that from Compiler2StreamCompiler.  (Note that the tuner apparently
doesn't use the default configuration -- we should add a custom technique
for that.)