Lower the max multiplier to Short.MAX_VALUE
Ideally we would set a bound on buffer size, then schedule the graph to
determine the max multiplier below that buffer size, but a) it's hard to
compute the size of Java objects, especially when only given their type
and b) the distributed compiler wants an all-reaching set of parameters
not dependent on the particular partitioning of the graph into blobs.

There are graphs on which one million might be a reasonable multiplier; if
single-digit items are buffered in the minimal schedule and the data is
ints or floats (4 bytes), one million items stays within some processors'
L3 caches.  But I want to constrain the tuner to more realistic numbers
for now (to improve tuning time, so I can test other tuning changes).
There's nothing special about Short.MAX_VALUE, it's just a convenient
non-random number.