IP address issue solved
In java, InetAddress.getLocalHost().getHostAddress() returns arbitrary IP
address based on how OS orders all available network interfaces in the system.
In Ubuntu, the system I tested, it often returns the loop-back address. So
the changes were made to keep track of actual IP addresses that the nodes
use to communicate with the controller. We need these actual IP address to
establish connection in between blobs.