TriangleContainment threaded implementation
I can't use ForkJoinPool here because we can't split an arbitrary Iterable
in half without materializing it into a List, which wouldn't be a fair
comparison.

The threaded implementation runs in about 16s on my Surface Pro 2,
compared to 27s for the StreamJIT implementation.  The StreamJIT
implementation is strangely insensitive to the multiplier, so perhaps
we're doing something wrong there.