Core: get smart about local storage
Previously we allocated way more local storage than our core actually
used, especially after fission.  This commit means we only allocate what
we actually use, which relieves pressure on the garbage collector now
that we aren't tying up lots of memory we don't use.

We could be even smarter and do "register allocation" on the local
storage to minimize our memory/cache footprint, but that's a lot of work
for probably not that much gain.