BlobHostStreamCompiler uses PermutationParameter for CPU affinity
We'll never pin two threads to the same core unless we're
oversubscribed, never pin one thread to more than one core, and never
leave a core unpinned.  We should also be trying to pin other JVM
threads, but there's no obvious way to do that on Linux because
processes and threads are both just tasks to the kernel.

The distributed runtime probably needs something more sophisticated than
this.