Sign error in FMRadio; StreamIt verified
I introduced a sign error when porting this benchmark to StreamJIT.  I
thought StreamIt was peeking because popping multiple items within a
statement is undefined (in C), but it's also to reverse their order.

With this, FMRadio 7, 128 generates the same output as StreamIt modulo
floating-point rounding and draining the graph (we do, StreamIt
doesn't).