Support 0 weights in WeightedRoundrobinSplitter
This can be used to fake multiple inputs, but note there must be at
least one real input for the stream to terminate.  It can also be used
to "inject" side inputs, such as key material from a key scheduling
filter in the crypto benchmarks.

Source filters used in the middle of the stream must still declare the
proper input type for their position in the graph so the compiler can
spin read method handles of the type it expects.  (The method handles
are never used, but they are still created.)

CheckVisitor will still complain, either about a source appearing in the
middle of the graph, or about unbalanced rates in the splitjoin (because
it multiplies the 0 into that branch's rate).