PeekableBuffer and required compiler changes
This allows "parallel input", by which I mean we don't copy input data
items into an array at the global barrier.  (It's not so much
parallelizing input as eliminating serial input.)

This only works when compiling the whole graph into a compiler blob due
to an ordering problem.  The compiler needs to know whether it's reading
from a PeekableBuffer or not during compilation.  Normally, a Blob only
learns about buffers when installBuffers is called, at which point the
Blob is already created.  And we can't just provide the buffers to
BlobFactory.makeBlob, because we need to know how big they should be,
which we don't learn until calling Blob.getMinimumBufferCapacity.  To
properly support this we'd need a two-phase compilation, where phase one
schedules/etc to compute buffer sizes, and phase two receives buffers
and produces the Blob.

When compiling the whole graph, only one blob needs the buffers, and we
can pass in Input (and Output) objects that can create buffers of any
required capacity, so we can side-step the issue at the cost of some
complexity.