Core.code: handle empty/duplicate internal storage
This fixes two different bugs.

The check for reads.isEmpty() avoids taking an exception there.  This
was uncovered by support for zero-rate splitters and joiners.

Moving the ActorGroup.specialize call into the loop avoids two
allocations both using one (possibly-too-small) ConcreteStorage when the
core has multiple allocations for the same ActorGroup.  This fix does
slightly complicate adding ActorGroup ordering parameters in the future,
but we'll cross that bridge when we come to it.