Compiler2 blobs handle short writes round-robin
This is similar to the interpreter problem from a while ago: we'll get
stuck waiting for space in one buffer while the downstream blob is
waiting for input on another buffer, causing a deadlock.

The interpreter has another issue if it has multiple sinks, and thus
must check what outputs are fireable before beginning.  The compiler
does not have this problem because it's scheduled -- all outputs execute
a fixed number of times.