Compiler2/tuner3: report throughput instead of execution time
When we time out the autotuner doesn't learn anything, but if we run
with a small enough input that we don't time out often, the fastest
configurations are very fast and susceptible to measurement noise.  Thus
we now report throughput for a fixed time period instead.  Note there's
still a compile timeout.