Compiler2/tuner3: broadcast allocation parameters to group
Currently we select one set of allocation parameters from each group and
ignore the rest.  When fusion changes, we select a different set of
parameters and lose our learning up to that point.

With these changes, we will now normalize our configurations such that
each group's allocation parameters are all set to the parameters
elected, so that small changes in fusion will still have access to the
learned values.

This doesn't address the issue where we want to split a group and give
it new allocation parameters at the same time, nor does it try to
average or otherwise merge allocation parameters when groups merge; thus
there's still some inertia to changing these parameters.

The implementation for this change is a total hack, but we'll figure out
something better later.