Compiler2: skip worker removal if we'd move initial data
When moving initial data, we need the moved data to appear before or
after the already-present initial data (splitter or joiner removal
respectively), but right now we don't adjust indices properly so we end
up overwriting the initial data.  But we still make liveness information
as if all data is present, so we take out-of-bounds or unboxing-null
exceptions as a result.

Sumanan and I worked out at the whiteboard that we could put the initial
data at negative indices and everything would fall into place, but too
many parts of the compiler now assume nonnegative indices for that to be
a prudent course of action at this time.  Moving the existing data is
difficult because the holes need to be at the front, not the back.

For now we'll assume initial data is on relatively few edges in the
graph and not worry about it too much.