Initial data on input or output prohibits removal
The exception is associated with an edge between two joiners.  This edge
has initial data so we can't remove the downstream joiner, but we do
remove the upstream one and propagate its inputSlots (liveness
information) downward.  That seems to mess things up, somehow.  As a
quick fix, checking both upstream and downstream storage for initial
data in findRemovals seems to work.  (Basically we don't care if it's a
splitter or a joiner; if there's initial data before or after it we
can't remove it.)