Index spans are defined by the first and last iteration only
This does assume the index functions are monotonically increasing with
respect to iteration (any decrease is strictly within an iteration).  This
passes all the tests, so it seems to be a safe assumption.

This significantly improves compile times at high multipliers, more from
avoiding lots of ImmutableSortedSet copies than from evaluating the index
functions fewer times.