Don't bytecodify in ActorGroup
This should never have been necessary.  I can only think of two reasons
we'd want to do this: to give each loop a meaningful class name for stack
traces, and to coax a reticent JVM into inlining.  On my Surface Pro 2,
this improves both compile and runtimes; we'll see if there's a runtime
impact when tuning it on the cluster.