Major IndexFunction changes to support bulk operations
Because paying one function call per index is not good enough at high
multipliers.

This is not fully implemented everywhere yet.

We may want to only operate on part of an array, so we don't have to have
an array of exactly the right size.

We probably want to tile loops in Compiler2 and elsewhere so we don't
trash the cache each time we do a bulk computation.