StorageSlot is one class with two distinguished instances
This means isLive and friends do not require virtual calls.  It seems to
be a minor win.

Unboxing StorageSlot is not obviously a win, because we need a long (8
bytes) per.  StorageSlot is 24 bytes but pointers to holes and dups are
only 4 bytes, so we might end up allocating more in total for graphs with
lots of duplicate splitters.